feat(ui): rename Settings block and expose item gap
