+++ packages/swingset/src/lib/registry.ts
@@ -107,12 +107,12 @@ import {
 } from '../stories/scroll-area.stories';
 import { meta as selectMeta } from '../stories/select.stories';
 import {
-  ConnectedAccounts as SettingsGroupConnectedAccounts,
-  Default as SettingsGroupDefault,
-  Destructive as SettingsGroupDestructive,
-  meta as settingsGroupMeta,
-  MultipleEmailAndPhoneNumbers as SettingsGroupMultipleEmailAndPhoneNumbers,
-} from '../stories/settings-group.stories';
+  ConnectedAccounts as SettingsConnectedAccounts,
+  Default as SettingsDefault,
+  Destructive as SettingsDestructive,
+  meta as settingsMeta,
+  MultipleEmailAndPhoneNumbers as SettingsMultipleEmailAndPhoneNumbers,
+} from '../stories/settings.stories';
 import { Default as TabsComponentDefault, meta as tabsComponentMeta } from '../stories/tabs.component.stories';
 import { meta as tabsMeta } from '../stories/tabs.stories';
 import {
@@ -133,12 +133,12 @@ import { toSlug } from './slug';
 import type { StoryModule } from './types';
 
 const destructiveModule: StoryModule = { meta: destructiveMeta, Default: DestructiveDefault };
-const settingsGroupModule: StoryModule = {
-  meta: settingsGroupMeta,
-  Default: SettingsGroupDefault,
-  MultipleEmailAndPhoneNumbers: SettingsGroupMultipleEmailAndPhoneNumbers,
-  ConnectedAccounts: SettingsGroupConnectedAccounts,
-  Destructive: SettingsGroupDestructive,
+const settingsModule: StoryModule = {
+  meta: settingsMeta,
+  Default: SettingsDefault,
+  MultipleEmailAndPhoneNumbers: SettingsMultipleEmailAndPhoneNumbers,
+  ConnectedAccounts: SettingsConnectedAccounts,
+  Destructive: SettingsDestructive,
 };
 const organizationProfileLeaveSectionModule: StoryModule = {
   meta: organizationProfileLeaveSectionMeta,
@@ -283,7 +283,7 @@ export const registry: StoryModule[] = [
   organizationProfileDeleteSectionModule,
   // Blocks
   destructiveModule,
-  settingsGroupModule,
+  settingsModule,
   // Components
   avatarModule,
   badgeModule,
