+++ packages/ui/src/mosaic/block/settings.vars.stylex.ts
@@ -1,7 +1,8 @@
 import * as stylex from '@stylexjs/stylex';
 
-import { colorVars } from '../tokens.stylex';
+import { colorVars, space } from '../tokens.stylex';
 
 export const settingsVars = stylex.defineVars({
   '--cl-settings-background': colorVars['--cl-color-background'],
+  '--cl-settings-items-gap': space['2'],
 });
